moved images to respective components
