+++ src/components/Footer/Footer.js
@@ -1,10 +1,10 @@
 import React from "react";
-import appStore from "../../images/appStore.svg";
-import facebook from "../../images/facebook.png";
-import googlePlay from "../../images/googlePlay.svg";
-import instagram from "../../images/instagram.png";
-import twitter from "../../images/twitter.png";
-import youtube from "../../images/youtube.png";
+import appStore from "./images/appStore.svg";
+import facebook from "./images/facebook.png";
+import googlePlay from "./images/googlePlay.svg";
+import instagram from "./images/instagram.png";
+import twitter from "./images/twitter.png";
+import youtube from "./images/youtube.png";
 
 export default function Footer(props) {
   return (
