+++ src/components/Header/HeaderLoggedIn/images/avatar.png
